Added business logic functionality of deleting customer selected form list-customers.jsp from database using hibernate

+++ WebContent/WEB-INF/view/list-customers.jsp
@@ -59,7 +59,7 @@
 							<!--  display the update and delete link --> 
 							<a href="${updateLink}">Update</a>
 							| 
-							<a href="${deleteLink}">Delete</a>
+							<a href="${deleteLink}" onclick="if(!(confirm('Are you sure you want to delete customer?'))) return false;">Delete</a>
 						</td>
 					</tr>
